requirements.txt
# Enhanced Deepfake Detection System Requirements
# ================================================

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0
tqdm>=4.66.0

# Database
# SQLite is built into Python

# Gemini API (optional)
google-generativeai>=0.3.0

# Configuration
pyyaml>=6.0

# Image hashing
imagehash>=4.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Dataset
kagglehub>=0.1.0

# Web Application (Flask)
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# Optional: Gradio for web interface
# gradio>=4.0.0
